Move same code to generic implementation.
