+++ src/main/scala/ch/becompany/social/github/GithubFeed.scala
@@ -2,26 +2,23 @@ package ch.becompany.social.github
 
 import java.time.Instant
 
-import akka.actor.ActorSystem
-import akka.pattern.after
-import akka.stream.scaladsl.Source
-import ch.becompany.social.{SocialFeed, Status}
+import ch.becompany.social.{PollingSocialFeed, Status}
 
 import scala.concurrent.duration._
 import scala.concurrent.{ExecutionContext, Future}
-import scala.util.{Failure, Try}
 
 /**
   * Streams GitHub events for an organization.
   * @param org The name of the organization.
   * @param ec The execution context.
   */
-class GithubFeed(org: String, userUpdateInterval: FiniteDuration = 1 minute)(implicit ec: ExecutionContext) extends SocialFeed {
+class GithubFeed(org: String, userUpdateInterval: FiniteDuration = 1 minute)(implicit ec: ExecutionContext)
+  extends PollingSocialFeed
+{
 
-  private val system = ActorSystem()
   // GitHub unauthenticated requests are limited to 60 per hour. => 1 req/m
   private val updateIntervalMin: FiniteDuration = 1 minute
-  private val updateInterval = updateIntervalMin.max(userUpdateInterval)
+  override val updateInterval = updateIntervalMin.max(userUpdateInterval)
 
   /**
     * Returns the latest `num` GitHub events.
@@ -30,36 +27,4 @@ class GithubFeed(org: String, userUpdateInterval: FiniteDuration = 1 minute)(imp
     */
   override def latest(num: Int): Future[List[(Instant, Status)]] =
     GithubClient.events(org).map(_.sortBy(_._1).takeRight(num))
-
-  private def events(lastUpdate: Instant): Future[List[(Instant, Try[Status])]] =
-    GithubClient.events(org).
-      map(_.
-        filter(_._1.isAfter(lastUpdate)).
-        sortBy(_._1).
-        map { case (date, status) => (date, Try(status)) }
-      ).
-      recover { case e => List(Instant.now -> Failure(e)) }
-
-  private def getLastUpdate(statuses: List[(Instant, Try[Status])]): Option[Instant] =
-    statuses.
-      lastOption.
-      map(_._1)
-
-  /**
-    * Streams GitHub events. The stream is populated using periodic requests;
-    * the interval between requests is set in the `updateInterval` value. The first
-    * request is emitted after `updateInterval`.
-    * @return The source providing the stream.
-    */
-  override def stream(): Source[(Instant, Try[Status]), _] =
-    Source.
-      unfoldAsync[Instant, List[(Instant, Try[Status])]](Instant.now) { lastUpdate =>
-        after(updateInterval, using = system.scheduler) {
-          events(lastUpdate) map { statuses =>
-            Some((getLastUpdate(statuses) getOrElse lastUpdate, statuses))
-          }
-        }
-      }.
-      flatMapConcat(list => Source.fromIterator(() => list.iterator))
-
 }
